fix: temporary fix for case issues
