+++ src/_components/MagicFontWeight.tsx
