+++ src/_components/thoughts/thoughts.css.ts
