+++ src/_components/window/image.tsx
